Some refactor, added plot functions and working CTrihedron.
plot2 and plot3 functionality for CBaseLidar to plot scanning rays.

CConfigLidar properties are set to protected so that they can only be change via constructor.
Angles, directions and lines are then set as properties (no methods).

CPlane3D is set to inherit from class CPose3D.
The points-in-plane checking is done with tolerance 1e-6 [m].

New class CPose3D with properties R, t and T and methods for plotting object frame and reference.

Inf intersection points are checked in CSimLidar for scanPolygon.
Angles, directions and lines from Lidar are refactored to access as properties.
plotPolygonScan is added to plot scanned points in 3D.

plotScan is added to CTrihedron to plot result of scan by Lidar